requirements.txt
pika == 1.2.1
pymongo == 4.1.1
python-dotenv == 0.20.0